chore: move scaleGasValue to its own util
